requirements.txt
pytest==5.1.2
selenium==3.141.0
allure-pytest==2.8.1
webdriver-manager==1.8.2
mimesis